AsserationError int ---> np.int64
Hi, 
I tried to use your project with OpenAi Gymnasium and Stable Baselines, but I got an error: 
"When using flat action space, action must be an Action object: 2 is invalid"
I printed the type of action_idx, and it was np.int64 I change assertion to numpy, and it works now.

+++ nasim/envs/action.py
@@ -695,7 +695,7 @@ class FlatActionSpace(spaces.Discrete):
         Action
             Corresponding Action object
         """
-        assert isinstance(action_idx, int), \
+        assert isinstance(action_idx, np.int64), \
             ("When using flat action space, action must be an integer"
              f" or an Action object: {action_idx} is invalid")
         return self.actions[action_idx]
